Implement user authentication checks in todos endpoint to ensure ownership validation

+++ TodoApp/routers/todos.py
@@ -10,7 +10,6 @@ from .auth import get_current_user
 
 router = APIRouter()
 
-
 def get_db():
   db = SessionLocal()
   try:
@@ -30,14 +29,19 @@ class TodoRequest(BaseModel):
      
 
 @router.get("/" ,  status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependancy):
-    todos = db.query(Todos).all()
-    return todos
+async def read_all(user: user_dependancy, db: db_dependancy):
+     if user is None:
+        raise HTTPException(status_code = 401 , details = 'Authentication Failed')
+     return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 @router.get("/todo/{todo_id}" , status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependancy, todo_id: int = Path(gt = 0) ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependancy, db: db_dependancy, todo_id: int = Path(gt = 0) ):
+    if user is None:
+        raise HTTPException(status_code = 401 , details = 'Authentication Failed')
+    todo_model = db.query(Todos).\
+    filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).first()
+
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
@@ -59,12 +63,17 @@ async def create_todo(user : user_dependancy,
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def update_todo(
+    user : user_dependancy,
     db: db_dependancy,
     todo_request: TodoRequest, 
     todo_id: int = Path(gt=0) 
     ):
 
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+      raise HTTPException(status_code = 401 , details = 'Authentication Failed')
+
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     
@@ -78,10 +87,15 @@ async def update_todo(
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependancy, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependancy, db: db_dependancy, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+
+    todo_model = db.query(Todos).\
+        filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).first()
+
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
-    
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+
+    db.query(Todos).filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).delete()
     db.commit()
